Bug Fixes/ImprovEMentsw
Fixed a bug preventing the broken ground to destroy itself when required

fixed a bug which if a earth construct is spawned before the enemies are spawned it would crash the game.
Also fixed issue where if the enemy it was chasing was destroyed it would break and remain afk.